Push vector layer and station loading down to map
